Day 21: Solved task 1

+++ aoc21/src/main/kotlin/de/havox_design/aoc2022/day21/Monkey.kt
@@ -1,10 +1,10 @@
 package de.havox_design.aoc2022.day21
 
-data class Monkey(var name: String, private var riddle: Riddle? = null, private var number: Int? = null) {
+data class Monkey(var name: String, private var riddle: Riddle? = null, private var number: Long? = null) {
     fun needsToSolveRiddle(): Boolean =
         number == null
 
-    fun calculateValue(monkeys: Collection<Monkey>): Int =
+    fun calculateValue(monkeys: Collection<Monkey>): Long =
         if (needsToSolveRiddle()) {
             riddle!!.calculate(monkeys)
         } else {
