+++ aoc21/src/main/kotlin/de/havox_design/aoc2022/day21/Riddle.kt
@@ -1,7 +1,7 @@
 package de.havox_design.aoc2022.day21
 
 data class Riddle(private var operandA: String, private var operation: Operator, private var operandB: String ) {
-    fun calculate(monkeys: Collection<Monkey>): Int {
+    fun calculate(monkeys: Collection<Monkey>): Long {
         val monkeyA = monkeys.first{monkey -> monkey.name == operandA }
         val monkeyB = monkeys.first{monkey -> monkey.name == operandB }
